`golang-debian` base image for building upstream toolchain (#667)
* Golang Debian base-image build

Kube cross, and other upstream tools, reley on a Debian-based Golang
image.

In order to allow for building and testing with the upstream toolchain
we are now building a `golang-debian` image using each available version
of EKS Go.

This commit adds a `golang-debian` build which builds a debian-based
golang base image; in pre-submit it just builds, in prod-post submit it
pushes to our public ECR for general consumption.

* remove golang req from RPM spec, as we dont install Golang in the RPM DB for the builder-base

* remove 'use_go_version' script as we're just setting the path directly
in the makefile

* merge env set